% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{getMeasure}
\alias{getMeasure}
\title{Extract measures from seed object}
\usage{
getMeasure(
  seed,
  measures = c("NPQ", "XE", "EF", "OE"),
  norm = c("local", "global", "none")
)
}
\arguments{
\item{seed}{a seed object}

\item{measures}{measure type to be extracted ("NPQ","XE","EF","OE")}

\item{norm}{how data should be normalised ("local","global","none")}
}
\value{
Seed object with extracted measures
}
\description{
Extract measures from seed object
}
