% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByIndiv.R
\name{getOrigin}
\alias{getOrigin}
\title{Extract measures from seed object}
\usage{
getOrigin(
  seed,
  splitby = c("plot", "pedigree", "line", "stem"),
  tagID = c("plot", "pedigree", "line", "stem")
)
}
\arguments{
\item{seed}{a seed object}

\item{splitby}{describes how data should be split ("plot","pedigree","line","stem")}

\item{tagID}{tags associated with mapping data}
}
\value{
Seed object with split data
}
\description{
Extract measures from seed object
}
