% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{getTraits}
\alias{getTraits}
\title{Extract traits from a seed object - simple trait and model traits}
\usage{
getTraits(seed, measure = c("NPQ", "XE", "EF", "OE"), cores = 1)
}
\arguments{
\item{seed}{a seed object}

\item{measure}{light measure that traits should be extracted for}

\item{cores}{numer of cores used for analysis}
}
\value{
return a seed object with extract traits
}
\description{
Extract traits from a seed object - simple trait and model traits
}
