% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{modelPlants}
\alias{modelPlants}
\title{Extract measures from seed object}
\usage{
modelPlants(
  seed,
  models = list(NPQ = c("beta", 3)),
  fit.to = c("plant", "allPlants", "medianPlant"),
  cores = 1
)
}
\arguments{
\item{seed}{a seed object}

\item{models}{list containing the name of the measure and the models to be run on data - Triple model only available for NPQ}

\item{fit.to}{describes how the data should be fitted ("plant","allPlants","medianPlant")}

\item{cores}{number of cores used for selecting plants}
}
\value{
Seed object with computed models
}
\description{
Extract measures from seed object
}
