% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportAndPlot.R
\name{plotSeed}
\alias{plotSeed}
\title{Plot data from seed object}
\usage{
plotSeed(seed, measure = c("NPQ", "EF", "XE", "OE"), dropped = FALSE)
}
\arguments{
\item{seed}{a seed object}

\item{measure}{light measure that should be used for plotting}

\item{dropped}{logical indicating if you wish to plot dropped samples}
}
\value{
plot with sample data -  may include models if they have been computed
}
\description{
Plot data from seed object
}
