% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{selectPlants}
\alias{selectPlants}
\title{Extract measures from seed object}
\usage{
selectPlants(
  seed,
  measure = c("NPQ", "XE", "EF", "OE"),
  method = c("MSE", "RSS"),
  threshold = c(0.05, 0.5),
  cores = 1
)
}
\arguments{
\item{seed}{a seed object}

\item{measure}{which measure should be used to select plants/samples ("NPQ","XE","EF","OE")}

\item{method}{if models have been computed, descirbes which method should be used to assess goodness of fit ("MSE","RSS")}

\item{threshold}{if models have been computed, threshold at which samples should be removed. Anything above threshold is removed.}

\item{cores}{number of cores used for selecting plants}
}
\value{
Seed object with filtered data
}
\description{
Extract measures from seed object
}
