% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{setStartPoints}
\alias{setStartPoints}
\title{Set Time points used in fluorImage analysis}
\usage{
setStartPoints(seed, start)
}
\arguments{
\item{seed}{A seed object}

\item{start}{a list containing starting values for gumbel optimisation. Please name your list elements a,b, and c}
}
\value{
returns a seed object with updated start points
}
\description{
Set Time points used in fluorImage analysis
}
