% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFunctions.R
\name{setTimePoints}
\alias{setTimePoints}
\title{Set Time points used in fluorImage analysis}
\usage{
setTimePoints(seed, timePoints)
}
\arguments{
\item{seed}{A seed object}

\item{timePoints}{a numeric vector containing end points of different light regimes}
}
\value{
returns a seed object with updated time points
}
\description{
Set Time points used in fluorImage analysis
}
