% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{sowSeed}
\alias{sowSeed}
\title{Create Seed object for lighteR analysis.}
\usage{
sowSeed(files, mapID = NULL, type = c("zone", "image"), areaThreshold = 5)
}
\arguments{
\item{files}{A directory containing Zone files and Image files}

\item{mapID}{A directory containing plant IDs}

\item{type}{Type of data that should be loaded - either zone and/or image}

\item{areaThreshold}{Area in square milimeters that should be considered as noise if appearing in grid}
}
\value{
Returns a new seed object - contains data provided in \code{files} and \code{mapID}
}
\description{
Create Seed object for lighteR analysis.
}
